fix: create useIsCurrentPage
