+++ src/lib/layout/navbar/index.tsx
@@ -1,10 +1,9 @@
 import { Flex } from "@chakra-ui/react";
 import { useWallet } from "@cosmos-kit/react";
 import { observer } from "mobx-react-lite";
-import { useRouter } from "next/router";
-import { useCallback } from "react";
 
 import { INSTANTIATED_LIST_NAME, getListIcon, SAVED_LIST_NAME } from "lib/data";
+import { useIsCurrentPage } from "lib/hooks/useIsCurrentPage";
 import { useContractStore, usePublicProjectStore } from "lib/providers/store";
 import { cmpContractListInfo } from "lib/stores/contract";
 import { formatSlugName } from "lib/utils";
@@ -22,29 +21,7 @@ const Navbar = observer(({ isExpand, setIsExpand }: NavbarProps) => {
   const { getContractLists } = useContractStore();
   const { getSavedPublicProjects } = usePublicProjectStore();
   const { currentChainRecord } = useWallet();
-
-  const router = useRouter();
-  const { network } = router.query;
-  const pathName = router.asPath;
-
-  const isCurrentPage = useCallback(
-    (slug: string) => {
-      const networkPath = network ? `/${network}` : "";
-      switch (slug) {
-        // handle home page
-        case "/":
-          return pathName === `${networkPath}` || pathName === "/";
-        // handle contract list page and public project page
-        case "/contract-list":
-        case "/public-project":
-          return pathName === `${networkPath}${slug}`;
-        // handle page with query param
-        default:
-          return pathName.includes(`${networkPath}${slug}`);
-      }
-    },
-    [network, pathName]
-  );
+  const isCurrentPage = useIsCurrentPage();
 
   const navMenu: MenuInfo[] = [
     {
